pyproject.toml
[project]
name = "django-authx"
version = "0.1.0"
description = "All in one authentication and authorization package for django"
authors = [
    { name = "DOMINIC MAABOBRA TUOLONG", email = "tuolongdominic@gmail.com" },
]

dependencies = ["django", "cryptography>=3.4", "pyotp>=2.9.0", "pyjwt>=2.10.1"]

requires-python = ">=3.10"
readme = "README.md"
license = { text = "BSD 3-Clause" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[project.optional-dependencies]
djangorestframework = ["djangorestframework", "django-filter", "markdown"]
phonenumbers = ["phonenumbers"]
all = ["django-authx[djangorestframework, phonenumbers]"]

[dependency-groups]
dev = ["pytest>=8.3.4", "pytest-django>=4.9.0"]
